Add HideFacetValueListener support to EDS backend.
