+++ config/vufind/EDS.ini
@@ -187,6 +187,16 @@ orFacets = *
 ; below, no facets will be translated.
 translated_facets[] = SourceType
 
+; Prevent specific facet values from being displayed to the user.
+; Use facet field names as keys and untranslated facet values as values.
+[HideFacetValue]
+;SourceType[] = "Book"
+
+; Prevent all but specific facet values from being displayed to the user.
+; Use facet field names as keys and untranslated facet values as values.
+[ShowFacetValue]
+;SourceType[] = "Book"
+
 ; This section controls the behavior of the EDS/Home screen.
 [HomePage]
 ; Content blocks can be selected from the list in searches.ini.
